+++ src/main/java/de/klinger/adw/repository/ClubRepository.java
@@ -1,13 +1,12 @@
 package de.klinger.adw.repository;
 
-import java.math.BigInteger;
 import java.util.List;
 
 import org.springframework.data.repository.CrudRepository;
 
 import de.klinger.adw.domain.Club;
 
-public interface ClubRepository extends CrudRepository<Club, BigInteger> {
+public interface ClubRepository extends CrudRepository<Club, Long> {
 
     @Override
     List<Club> findAll();
